+++ readium/opds/r2-opds/src/test/java/org/readium/r2/opds/TestUtils.kt
@@ -12,8 +12,6 @@ package org.readium.r2.opds
 import org.json.JSONArray
 import org.json.JSONObject
 import org.junit.Assert
-import org.readium.r2.shared.extensions.toList
-import org.readium.r2.shared.extensions.toMap
 import java.io.File
 import java.net.URL
 
@@ -21,14 +19,14 @@ import java.net.URL
  * Asserts that two [JSONObject] are equal.
  */
 fun assertJSONEquals(expected: JSONObject, actual: JSONObject) {
-    Assert.assertEquals(expected.toMap(), actual.toMap())
+    Assert.assertEquals(expected.toString(), actual.toString())
 }
 
 /**
  * Asserts that two [JSONArray] are equal.
  */
 fun assertJSONEquals(expected: JSONArray, actual: JSONArray) {
-    Assert.assertEquals(expected.toList(), actual.toList())
+    Assert.assertEquals(expected.toString(), actual.toString())
 }
 
 class Fixtures(val path: String? = null) {