+++ readium/opds/CHANGELOG.md
@@ -10,6 +10,10 @@ All notable changes to this project will be documented in this file.
 
 * New APIs using coroutines and R2's `HttpClient` instead of Fuel and kovenant (contributed by [@stevenzeck](https://github.com/readium/r2-opds-kotlin/pull/55)).
 
+### Changed
+
+* Upgraded to Kotlin 1.5.20.
+
 ### Fixed
 
 * Links in an OPDS 2 feed are normalized to the feed base URL.